% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colp_search.R
\name{colp_search}
\alias{colp_search}
\title{Search Catalogue of Life Plus}
\usage{
colp_search(
  q = NULL,
  id = NULL,
  datasetKey = NULL,
  rank = NULL,
  nomstatus = NULL,
  status = NULL,
  issue = NULL,
  type = NULL,
  publishedIn = NULL,
  hasField = NULL,
  facet = NULL,
  sortBy = NULL,
  start = 0,
  limit = 10,
  ...
)
}
\arguments{
\item{q}{(character) main query string}

\item{id}{(character) the name identifier}

\item{datasetKey}{(character) dataset key}

\item{rank}{(character) filter by rank. one of: domain, superkingdom,
kingdom, subkingdom, infrakingdom, superphylum, phylum, subphylum,
infraphylum, superclass, class, subclass, infraclass, parvclass,
superlegion, legion, sublegion, infralegion, supercohort, cohort,
subcohort, infracohort, magnorder, superorder, grandorder, order,
suborder, infraorder, parvorder, superfamily, family, subfamily,
infrafamily, supertribe, tribe, subtribe, infratribe, suprageneric name,
genus, subgenus, infragenus, supersection, section, subsection, superseries,
series, subseries, infrageneric name, species aggregate, species,
infraspecific name, grex, subspecies, cultivar group, convariety,
infrasubspecific name, proles, natio, aberration, morph, variety,
subvariety, form, subform, pathovar, biovar, chemovar, morphovar, phagovar,
serovar, chemoform, forma specialis, cultivar, strain, other, unranked}

\item{nomstatus}{(character) filter by nomenclatural status. one of: ok,
unavailable, illegitimate, variant, conserved, rejected, doubtful, unevaluated}

\item{status}{(character) filter by taxonomic status. one of: accepted,
doubtful, ambiguous synonym}

\item{issue}{(character) filter by issue found}

\item{type}{(character) filter by name type. one of: scientific, virus,
hybrid-formula, cultivar, otu, placeholder, none}

\item{publishedIn}{(character) reference id to filter names by}

\item{hasField}{(logical) filter that only includes name where the
requested name property is not \code{NULL}}

\item{facet}{(character) request a facet to be returned. one of:
dataset_key, rank, nom_status, status, issue, type, field. facet
limit default: 50}

\item{sortBy}{(character) one of: "relevance", "name", "key"}

\item{start}{(integer) requested number of offset records. Default: 0}

\item{limit}{(integer) requested number of maximum records to be returned.
Default: 10; max: 1000}

\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\description{
Search Catalogue of Life Plus
}
\examples{
\dontrun{
colp_search(q="Apis")
colp_search(q="Agapostemon")
colp_search(q="Agapostemon", datasetKey = 3)
colp_search(q="Agapostemon", rank = "genus")
colp_search(q="Agapostemon", nomstatus = "doubtful")
colp_search(q="Agapostemon", status = "accepted")
colp_search(type = "virus")
colp_search(q="Bombus", facet = "rank")

x <- colp_search(q="Poa")
x
x$result
x$result$usage
x$result$usage$name
}
}
