% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colpluz-package.R
\docType{package}
\name{colpluz-package}
\alias{colpluz-package}
\alias{colpluz}
\title{colpluz}
\description{
Catalogue of Life Plus Client
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
