% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_classification.R
\name{cp_classification}
\alias{cp_classification}
\title{Classification}
\usage{
cp_classification(dataset_key, taxon_id, ...)
}
\arguments{
\item{dataset_key}{(character) dataset identifier}

\item{taxon_id}{(character) taxon identifier}

\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\description{
Classification
}
\examples{
\dontrun{
z <- cp_classification(dataset_key=1000, taxon_id='10')
z
cp_classification(dataset_key=1000, taxon_id='20')
cp_classification(dataset_key=3,
 taxon_id='709dfa66-43fe-4736-a774-59f821a2ee23')
}
}
