% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_decision.R
\name{cp_decision}
\alias{cp_decision}
\title{Descisions}
\usage{
cp_decision(
  dataset_key = NULL,
  subject_dataset_key = NULL,
  user_key = NULL,
  rank = NULL,
  mode = NULL,
  broken = NULL,
  start = 0,
  limit = 10,
  ...
)
}
\arguments{
\item{dataset_key}{(character) The catalogue dataset the sector is attached
to. optional}

\item{subject_dataset_key}{(character) The source dataset the subject
belongs to. optional}

\item{user_key}{(character) The user that last modified the sector. optional}

\item{rank}{(character) The rank of the subject. optional. one of the
rank types:
https://catalogueoflife.github.io/general/api/api.html#model-RankType}

\item{mode}{(character) The sector mode, e.g. ATTACH. optional}

\item{broken}{(logical) If \code{TRUE} only show decisions which cannot be linked
to a source name usage. optional}

\item{start}{(integer) requested number of offset records. Default: 0}

\item{limit}{(integer) requested number of maximum records to be returned.
Default: 10; max: 1000}

\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}

\item{(character)}{a dataset key to filter by. optional}
}
\value{
list
}
\description{
List of editorial decisions with optional filters.
}
\examples{
\dontrun{
x <- cp_decision()
x
}
}
