% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_nameusage_suggest.R
\name{cp_nameusage_suggest}
\alias{cp_nameusage_suggest}
\title{Name Usage: Suggest}
\usage{
cp_nameusage_suggest(q, datasetKey, vernaculars = FALSE, limit = 10, ...)
}
\arguments{
\item{q}{(character) main query string. required}

\item{datasetKey}{(character) dataset key. required}

\item{vernaculars}{(logical) Whether or not to include vernacular names
in the suggestions (default: \code{FALSE})}

\item{limit}{(integer) requested number of maximum records to be returned.
Default: 10; max: 1000}

\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
a tibble
}
\description{
Name Usage: Suggest
}
\examples{
\dontrun{
cp_nameusage_suggest(q="Apis", datasetKey = 3)
}
}
