% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_parser.R
\name{cp_parser}
\alias{cp_parser}
\title{Name Parser}
\usage{
cp_parser(names, ...)
}
\arguments{
\item{names}{(character) one or more scientific names to parse}

\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
tibble, with one row for each parsed name
}
\description{
Name Parser
}
\examples{
\dontrun{
cp_parser(names = "Apis mellifera")
cp_parser(names = c("Apis mellifera", "Homo sapiens var. sapiens"))
}
}
