% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_sector.R
\name{cp_sector}
\alias{cp_sector}
\title{Sectors}
\usage{
cp_sector(
  id = NULL,
  dataset_key = NULL,
  subject_dataset_key = NULL,
  user_key = NULL,
  rank = NULL,
  mode = NULL,
  broken = NULL,
  start = 0,
  limit = 10,
  ...
)
}
\arguments{
\item{id}{(character) The subjects taxon id. optional}

\item{dataset_key}{(character) The catalogue dataset the sector is attached
to. optional}

\item{subject_dataset_key}{(character) The source dataset the subject
belongs to. optional}

\item{user_key}{(character) The user that last modified the sector. optional}

\item{rank}{(character) The rank of the subject. optional. one of the
rank types:
https://catalogueoflife.github.io/general/api/api.html#model-RankType}

\item{mode}{(character) The sector mode, e.g. ATTACH. optional}

\item{broken}{(logical) If \code{TRUE} only show decisions which cannot be linked
to a source name usage. optional}

\item{start}{(integer) requested number of offset records. Default: 0}

\item{limit}{(integer) requested number of maximum records to be returned.
Default: 10; max: 1000}

\item{...}{curl options passed on to \code{\link[crul]{verb-GET}}}
}
\value{
list
}
\description{
Sectors are taxonomic groups that are represented by a subtree with a
single root taxon. They link a subject (the root taxon in a source dataset)
to a target (the attachment point) in a managed catalogue.
}
\examples{
\dontrun{
cp_sector()
}
}
