% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{cp_up}
\alias{cp_up}
\title{check if api up in examples}
\usage{
cp_up(x)
}
\value{
boolean. \code{TRUE} if http request succeeds, \code{FALSE} if not
}
\description{
check if api up in examples
}
\keyword{internal}
