% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_vocab.R
\name{cp_vocab}
\alias{cp_vocab}
\title{CoL Vocabularies}
\usage{
cp_vocab(vocab, ...)
}
\arguments{
\item{vocab}{(character) a vocabulary name}

\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
character vector of words
}
\description{
CoL Vocabularies
}
\examples{
\dontrun{
cp_vocab("rank")
cp_vocab("datasetorigin")
cp_vocab("datasettype")
cp_vocab("matchtype")
cp_vocab("taxonomicstatus")
}
}
