% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_ds.R
\name{cp_ds}
\alias{cp_ds}
\title{Datasets API route catch all method}
\usage{
cp_ds(route, ...)
}
\arguments{
\item{route}{(character) an API route. the \verb{/dataset} route
part is added internally; so just include the route following that.
required.}

\item{...}{named parameters, passed on to \code{\link[glue:glue]{glue::glue()}}. required.
param names must match \verb{\\\\\{key\\\\\}} given in the \code{route}}
}
\description{
Datasets API route catch all method
}
\details{
There are A LOT of datasets API routes. Instead of making
an R function for each route, we have R functions for some of the
"more important" routes, then \code{cp_ds()} will allow you to make
requests to the remainder of the datasets API routes.
}
\section{Not supported dataset routes}{

Some dataset routes do not return JSON so we don't support those.
Thus far, the only route we don't support is \verb{/dataset/\\\\\{key\\\\\}/logo}
}

\examples{
cp_ds(route = "{key}/namesdiff", key = "1000")
cp_ds(route = "{key}/reference", key = "1000")
cp_ds(route = "{key}/tree", key = "1000")
cp_ds(route = "{key}/tree", key = "1014")
}
