% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_nameusage_suggest.R
\name{cp_nameusage_suggest}
\alias{cp_nameusage_suggest}
\title{Name Usage: Suggest}
\usage{
cp_nameusage_suggest(
  q = NULL,
  datasetKey = NULL,
  vernaculars = FALSE,
  limit = 10,
  ...
)
}
\arguments{
\item{q}{(character) main query string}

\item{datasetKey}{(character) dataset key}

\item{vernaculars}{(logical) Whether or not to include vernacular names
in the suggestions (default: \code{FALSE})}

\item{limit}{(integer) requested number of maximum records to be returned.
Default: 10; max: 1000}

\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\description{
Name Usage: Suggest
}
