% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcol-package.R
\docType{package}
\name{rcol-package}
\alias{rcol}
\alias{rcol-package}
\title{rcol: R Client for the Catalogue of Life / ChecklistBank API}
\description{
Access the Catalogue of Life (COL) through the ChecklistBank (CLB) web API.
Match scientific names against any public dataset or the latest COL release,
look up datasets, taxa and name usages, navigate the taxonomic tree, run the
ChecklistBank name and value parsers, and discover the latest annual and
monthly COL base and extended releases.
}
\seealso{
Useful links:
\itemize{
\item \url{https://github.com/CatalogueOfLife/rcol}
\item \url{https://catalogueoflife.github.io/rcol/}
\item Report bugs at \url{https://github.com/CatalogueOfLife/rcol/issues}
}
}
\author{
\strong{Maintainer}: Markus Döring \email{mdoering@gbif.org} (\href{https://orcid.org/0000-0001-7757-1889}{ORCID})

Other contributors:
\itemize{
\item Catalogue of Life [copyright holder, funder]
}
}
\keyword{internal}
