% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcol-package.R
\docType{package}
\name{rcol-package}
\alias{rcol-package}
\alias{rcol}
\title{rcol}
\description{
Catalogue of Life Plus (CoL) Client
}
\note{
CoL API docs are at https://api.catalogueoflife.org/
}
\author{
Scott Chamberlain
}
\keyword{package}
